add gitignore, add hourly data
