+++ .gitignore
@@ -1,0 +1,5 @@
+.idea/
+__pycache__/
+config.py
+data.py
+venv/
